# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
    AddAttachmentRequest:
      type: object
      required:
      - label
      - source
      - dataBase64
      properties:
        mimeType:
          type: string
        label:
          type: string
          pattern: \S
        source:
          type: string
          pattern: \S
        dataBase64:
          type: string
          pattern: \S
    AgentActivityState:
      type: string
      enum:
      - IDLE
      - BUSY
      - WAITING
      - ENDED
    ApiError:
      type: object
      description: "The error envelope: a message, plus a reason on a 409."
      properties:
        message:
          type: string
          description: "What went wrong, in words. Always present."
        reason:
          $ref: "#/components/schemas/Reason"
          type: string
          description: "Which refusal this is. Present on the 409s the release and\
            \ integrate flows raise, absent on every other error. A value a client\
            \ does not recognise is shown as the message verbatim."
        conflicts:
          type: array
          items:
            type: string
          description: The conflicted paths. Present only for CONFLICT and MERGE_CONFLICT.
...
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
...
    PromptAttachmentSource:
      type: string
      enum:
      - SKETCH
      - PASTE
    Reason:
      type: string
      enum:
      - CONFLICT
      - MERGE_CONFLICT
      - NOT_FAST_FORWARD
      - ALREADY_INTEGRATED
      - PUSH_REJECTED
      - VERSION_ALREADY_RELEASED
      - RELEASE_REQUIRED
...
    Response17:
      type: object
      properties:
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/WorkspacePromptAttachmentDataDto"
    Response18:
      type: object
      properties:
        id:
          type: string
        mimeType:
          type: string
        label:
          type: string
        source:
          $ref: "#/components/schemas/PromptAttachmentSource"
        createdAt:
          $ref: "#/components/schemas/Instant"
    Response19:
      type: object
      properties:
        draft:
          $ref: "#/components/schemas/WorkspacePromptDraftDto"
...
    Response20:
      type: object
      properties:
        draft:
          $ref: "#/components/schemas/WorkspacePromptDraftDto"
...
    Response9:
      type: object
      properties:
        workspace:
          $ref: "#/components/schemas/WorkspaceDto"
    SavePromptDraftRequest:
      type: object
      required:
      - content
      properties:
        content:
          type: string
        serializedPrompt:
          type: string
...
    WorkspaceDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        workspaceId:
          type: string
        parent:
          type: string
        branch:
          type: string
        ahead:
          type: integer
          format: int32
        behind:
          type: integer
          format: int32
        conflictsWithParent:
          type: boolean
        status:
          $ref: "#/components/schemas/WorkspaceStatus"
        runtimeStatus:
          $ref: "#/components/schemas/WorkspaceRuntimeStatus"
        runtimeError:
          type: string
        clean:
          type: boolean
        agentActivity:
          $ref: "#/components/schemas/AgentActivityState"
        preamble:
          type: string
        result:
          type: string
        resolvedAt:
          $ref: "#/components/schemas/Instant"
        daemonConnectedAt:
          $ref: "#/components/schemas/Instant"
        daemonVersion:
          type: string
        daemonBuildTime:
          $ref: "#/components/schemas/Instant"
        daemonOutdated:
          type: boolean
...
    WorkspacePromptAttachmentDataDto:
      type: object
      properties:
        id:
          type: string
        mimeType:
          type: string
        label:
          type: string
        source:
          $ref: "#/components/schemas/PromptAttachmentSource"
        createdAt:
          $ref: "#/components/schemas/Instant"
        dataBase64:
          type: string
    WorkspacePromptDraftDto:
      type: object
      properties:
        content:
          type: string
        serializedPrompt:
          type: string
        promptVersion:
          type: integer
          format: int64
        lastRunAt:
          $ref: "#/components/schemas/Instant"
        lastRunPromptVersion:
          type: integer
          format: int64
        lastRunCommandId:
          type: string
        updatedAt:
          $ref: "#/components/schemas/Instant"
    WorkspaceRuntimeStatus:
      type: string
      enum:
      - RUNNING
      - STOPPED
      - PROVISIONING
      - FAILED
    WorkspaceStatus:
      type: string
      enum:
      - ACTIVE
      - INTEGRATED
      - ABANDONED
paths:
...
  /workspaces/api/workspaces/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: The workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response9"
        "404":
          description: No such ACTIVE workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
      summary: Get
      tags:
      - Workspace Controller
...
  /workspaces/api/workspaces/{id}/prompt-attachments:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: "The attachments, oldest first."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response17"
        "404":
          description: No such ACTIVE workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
      summary: List
      tags:
      - Workspace Prompt Attachment Controller
    post:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddAttachmentRequest"
        required: true
      responses:
        "201":
          description: Attached; the row without its bytes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response18"
        "400":
          description: "Not valid base64, not a PNG or JPEG, or an unknown source."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: No such ACTIVE workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "413":
          description: The image exceeds the configured per-image byte cap.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
      summary: Add
      tags:
      - Workspace Prompt Attachment Controller
  /workspaces/api/workspaces/{id}/prompt-attachments/{attachmentId}:
    delete:
      parameters:
      - name: attachmentId
        in: path
        required: true
        schema:
          type: string
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Removed.
        "404":
          description: "No such ACTIVE workspace, or no such attachment on it."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
      summary: Delete
      tags:
      - Workspace Prompt Attachment Controller
  /workspaces/api/workspaces/{id}/prompt-draft:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavePromptDraftRequest"
        required: true
      responses:
        "200":
          description: "The persisted draft, with its fresh updatedAt."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response19"
        "400":
          description: The content is not a well-formed JSON document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: No such ACTIVE workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "413":
          description: The draft exceeds the configured byte cap.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
      summary: Save
      tags:
      - Workspace Prompt Draft Controller
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: The saved draft.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response20"
        "404":
          description: "No such ACTIVE workspace, or it has no saved draft."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
      summary: Get
      tags:
      - Workspace Prompt Draft Controller
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: There is no draft for this workspace.
        "404":
          description: No such ACTIVE workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
      summary: Delete
      tags:
      - Workspace Prompt Draft Controller
...
info:
  title: qits-workspaces API
  version: 1.0.0-SNAPSHOT